product_service/requirements.txt
aws-cdk-lib==2.145.0
constructs>=10.0.0,<11.0.0
boto3==1.34.131
Faker==25.9.1